Move reducer to a separate directory + remove unused imports/fix issue with onClick on similar_photos for the modal view
