+++ frontend/src/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import './App.scss';
-import HomeRoute from './components/HomeRoute';
+import HomeRoute from './routes/HomeRoute';
 import PhotoDetailsModal from 'routes/PhotoDetailsModal';
 import useApplicationData from 'hooks/useApplicationData';
