+++ frontend/src/components/TopicList.jsx
@@ -3,29 +3,11 @@ import React from "react";
 import "../styles/TopicList.scss";
 import TopicListItem from "./TopicListItem";
 
-const sampleDataForTopicList = [
-  {
-    id: "1",
-    slug: "topic-1",
-    title: "Nature",
-  },
-  {
-    id: "2",
-    slug: "topic-2",
-    title: "Travel",
-  },
-  {
-    id: "3",
-    slug: "topic-3",
-    title: "People",
-  },
-];
-
 const TopicList = (props) => {
-  const topics = props.topics.map(({id, slug, title}, i) => {
+  const topics = props.topics.map(({id, slug, title}) => {
     return (
       <TopicListItem
-        key={i}
+        key={id}
         id={id}
         slug={slug}
         title={title}
