+++ frontend/src/components/TopicListItem.jsx
@@ -2,12 +2,6 @@ import React from "react";
 
 import "../styles/TopicListItem.scss";
 
-const sampleDataForTopicListItem = {
-  id: "1",
-  slug: "topic-1",
-  label: "Nature",
-};
-
 const TopicListItem = (props) => {
   const {id, slug, title} = props;
   return (
