+++ frontend/src/hooks/useApplicationData.js
@@ -1,55 +1,5 @@
 import React, { useEffect, useReducer, useCallback } from 'react';
-
-export const ACTIONS = {
-  FAV_PHOTO_ADDED: 'FAV_PHOTO_ADDED',
-  FAV_PHOTO_REMOVED: 'FAV_PHOTO_REMOVED',
-  FAV_ICON_LIKED: 'FAV_ICON_LIKED',
-  FAV_ICON_UNLIKED: 'FAV_ICON_UNLIKED',
-  SET_PHOTO_DATA: 'SET_PHOTO_DATA',
-  SET_TOPIC_DATA: 'SET_TOPIC_DATA',
-  SELECT_PHOTO: 'SELECT_PHOTO',
-  DISPLAY_PHOTO_DETAILS: 'DISPLAY_PHOTO_DETAILS',
-  OPEN_MODAL: 'OPEN_MODAL',
-  CLOSE_MODAL: 'CLOSE_MODAL',
-  ADD_FAV_NOTIFICATION: 'ADD_FAV_NOTIFICATION',
-  REMOVE_FAV_NOTIFICATION: 'REMOVE_FAV_NOTIFICATION',
-  SET_PHOTOS_BY_TOPIC: 'SET_PHOTOS_BY_TOPIC',
-  SET_CURRENT_TOPIC: 'SET_CURRENT_TOPIC'
-};
-
-const reducer = (state, action) => {
-  switch (action.type) {
-  case ACTIONS.OPEN_MODAL: {
-    return { ...state, isModalOpen: true, selectedPhoto: action.payload };
-  }
-  case ACTIONS.CLOSE_MODAL: {
-    return { ...state, isModalOpen: false, selectedPhoto: null };
-  }
-  case ACTIONS.FAV_PHOTO_ADDED: {
-    return { ...state, favorites: new Set([...state.favorites, action.payload]) };
-  }
-  case ACTIONS.FAV_PHOTO_REMOVED: {
-    return {...state, favorites: new Set([...state.favorites].filter(item => item !== action.payload)) };
-  }
-  case ACTIONS.ADD_FAV_NOTIFICATION: {
-    return {...state, displayAlert: state.favorites && state.favorites.size > 0 };
-  }
-  case ACTIONS.SET_PHOTO_DATA: {
-    return {...state, photos: action.payload };
-  }
-  case ACTIONS.SET_TOPIC_DATA: {
-    return {...state, topics: action.payload };
-  }
-
-  case ACTIONS.SET_CURRENT_TOPIC: {
-    return {...state, currentTopic: action.payload };
-  }
-
-  default: {
-    return state;
-  }
-  }
-};
+import reducer, { ACTIONS } from '../reducers/reducer';
 
 const useApplicationData = () => {
   const [state, dispatch] = useReducer(reducer, {
